+++ publish/relaychat/relaychat_test.go
@@ -5,13 +5,29 @@
 package relaychat // import "github.com/wabarc/wayback/publish/relaychat"
 
 import (
-	"os"
 	"testing"
+
+	"github.com/wabarc/helper"
+	"github.com/wabarc/wayback/config"
 )
 
-func setIRCEnv() {
-	os.Setenv("WAYBACK_IRC_NICK", "foo")
-	os.Setenv("WAYBACK_IRC_CHANNEL", "bar")
+const server = "irc.libera.chat:6697"
+
+func setIRCEnv(t *testing.T) {
+	t.Setenv("WAYBACK_IRC_NICK", helper.RandString(6, ""))
+	t.Setenv("WAYBACK_IRC_CHANNEL", "bar")
+	t.Setenv("WAYBACK_IRC_SERVER", server)
 }
 func TestToIRCChannel(t *testing.T) {
+}
+
+func TestShutdown(t *testing.T) {
+	setIRCEnv(t)
+	opts, _ := config.NewParser().ParseEnvironmentVariables()
+
+	irc := New(nil, opts)
+	err := irc.Shutdown()
+	if err != nil {
+		t.Errorf("Unexpected shutdown: %v", err)
+	}
 }