+++ publish/telegram/telegram.go
@@ -119,3 +119,10 @@ func (t *Telegram) toChannel(art reduxer.Artifact, head, body string) (ok bool) 
 
 	return true
 }
+
+// Shutdown shuts down the Telegram service.
+func (t *Telegram) Shutdown() error {
+	t.bot.Stop()
+
+	return nil
+}