+++ publish/twitter/twitter_test.go
@@ -8,7 +8,6 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"os"
 	"strings"
 	"testing"
 
@@ -18,15 +17,15 @@ import (
 	"github.com/wabarc/wayback/template/render"
 )
 
-func setTwitterEnv() {
-	os.Setenv("WAYBACK_TWITTER_CONSUMER_KEY", "foo")
-	os.Setenv("WAYBACK_TWITTER_CONSUMER_SECRET", "foo")
-	os.Setenv("WAYBACK_TWITTER_ACCESS_TOKEN", "foo")
-	os.Setenv("WAYBACK_TWITTER_ACCESS_SECRET", "foo")
+func setTwitterEnv(t *testing.T) {
+	t.Setenv("WAYBACK_TWITTER_CONSUMER_KEY", "foo")
+	t.Setenv("WAYBACK_TWITTER_CONSUMER_SECRET", "foo")
+	t.Setenv("WAYBACK_TWITTER_ACCESS_TOKEN", "foo")
+	t.Setenv("WAYBACK_TWITTER_ACCESS_SECRET", "foo")
 }
 
 func TestToTwitter(t *testing.T) {
-	setTwitterEnv()
+	setTwitterEnv(t)
 
 	client, mux, server := helper.MockServer()
 	defer server.Close()
@@ -52,5 +51,20 @@ func TestToTwitter(t *testing.T) {
 	got := twitt.ToTwitter(context.Background(), txt)
 	if !got {
 		t.Errorf("Unexpected create GitHub Issues got %t instead of %t", got, true)
+	}
+}
+
+func TestShutdown(t *testing.T) {
+	setTwitterEnv(t)
+
+	opts, _ := config.NewParser().ParseEnvironmentVariables()
+
+	httpClient, _, server := helper.MockServer()
+	defer server.Close()
+
+	tw := New(httpClient, opts)
+	err := tw.Shutdown()
+	if err != nil {
+		t.Errorf("Unexpected shutdown: %v", err)
 	}
 }