+++ docs/changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Set NoWait to true for tor listen config ([#368](https://github.com/wabarc/wayback/pull/368))
 - Improve context deadline ([#381](https://github.com/wabarc/wayback/pull/381))
+- Improve publish services ([#382](https://github.com/wabarc/wayback/pull/382))
 
 ## [0.19.1] - 2023-03-21